+++ spec/js_routes/options_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+require 'byebug'
 
 describe JsRoutes, "options" do
 
@@ -519,6 +520,15 @@ describe JsRoutes, "options" do
     it "disables documentation generation" do
       expect(generated_js).not_to include("@param")
       expect(generated_js).not_to include("@returns")
+    end
+  end
+
+  describe "module.exports usage" do
+    it "doesn't include module.exports" do
+      # We mention `module.exports` in the comments of routes.ts, just so it's clear
+      # why the expression is so weird there.
+      js_without_comments = generated_js.lines.map { it.gsub(/\/\/.*$/, "") }.join("\n")
+      expect(js_without_comments).not_to include("module.exports")
     end
   end